msistoreapp/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.36.0
cryptography==41.0.5
Deprecated==1.2.14
Django==4.2.7
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
wrapt==1.15.0
